% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsdrtools-classes.R
\name{as_tsdrtools_install}
\alias{as_tsdrtools_install}
\title{Create tsdrtools_install}
\usage{
as_tsdrtools_install(x)
}
\arguments{
\item{x}{list to make tsdrtools_install}
}
\description{
Create tsdrtools_install
}
