% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsdrtools-classes.R
\name{as_tsdrtools_status}
\alias{as_tsdrtools_status}
\title{Create tsdrtools_status}
\usage{
as_tsdrtools_status(x)
}
\arguments{
\item{x}{list to make tsdrtools_status}
}
\description{
Create tsdrtools_status
}
