% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsdrtools-classes.R
\name{error_logs}
\alias{error_logs}
\title{Extract error logs from tsdrstools_install-object}
\usage{
error_logs(x, package = NULL)
}
\arguments{
\item{x}{object of class tsdrstools_install}

\item{package}{NULL or character vector of package errors to extract. If NULL extracts all.}
}
\value{
list of errors
}
\description{
Extract error logs from tsdrstools_install-object
}
\examples{
\dontrun{
status <- tsd_package_install("devtools.zip")
error_logs(status)
}
}
