% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_dependency_tree}
\alias{get_dependency_tree}
\title{Get package dependency tree}
\usage{
get_dependency_tree(
  package,
  i = 1,
  type = c("Depends", "Imports"),
  repos = getOption("repos"),
  verbose = TRUE,
  ...
)
}
\arguments{
\item{package}{name of package. character.}

\item{i}{integer. Used in recursive call to package dependency retrieval.}

\item{type}{Character vector, a subset of c("Depends", "Imports", "LinkingTo", "Suggests", "Enhances")}

\item{repos}{character vector. Addresses of CRAN mirror(s)}

\item{verbose}{logical. If function should be talkative.}

\item{...}{additional arguments to functions \code{\link[utils]{available.packages}}}
}
\value{
character vector of unique packages
}
\description{
This function will return all package dependencies
for any given package, including backwards dependency
for all.
The function mainly prints out the dependency tree
in the console, but also returns all packages
required for build (uniquely).
}
