% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsdrtools-classes.R
\name{install_error}
\alias{install_error}
\title{Constructor for install_error}
\usage{
install_error(x)
}
\arguments{
\item{x}{named list of installation errors}
}
\description{
Constructor for install_error
}
