% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsdrtools-classes.R
\name{is_tsdrtools_install}
\alias{is_tsdrtools_install}
\title{Validate tsdrtools_install}
\usage{
is_tsdrtools_install(x)
}
\arguments{
\item{x}{an object}
}
\description{
Validate tsdrtools_install
}
