% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg-workarounds.R
\name{package_workarounds}
\alias{package_workarounds}
\title{Retrieve package workaround}
\usage{
package_workarounds(package, folder, type, verbose = TRUE)
}
\arguments{
\item{package}{name of package. character.}

\item{folder}{character. folder to place all necessary files in}

\item{type}{character vector of either "prepare" or "install".
This indicates if the workaround for install-process or
prepare-process should be returned.}

\item{verbose}{logical. If function should be talkative.}
}
\description{
Certain R-packages require workarounds
to install on offline servers.
This function will check if the given package
has such a work-around and return said
workaround for the type of process asked
(prepare or install).
}
