% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg-workarounds.R
\name{stringi_workaround}
\alias{stringi_workaround}
\title{Function for stringi workaround}
\usage{
stringi_workaround(folder, type, verbose = TRUE)
}
\arguments{
\item{folder}{character. folder to place all necessary files in}

\item{type}{character vector of either "prepare" or "install".
This indicates if the workaround for install-process or
prepare-process should be returned.}

\item{verbose}{logical. If function should be talkative.}
}
\description{
stringi requires download of a bundle
to work. If we want to install offline,
the bundle needs to be downloaded and
ported with the package tar.
For \code{type = "prepare"} the bundle is
downloaded into the folder with the pacakge
tars.
For install, the call to \code{R CMD INSTALL}
is given the path to the bundle during installation,
which will stop it from attempting to
download the bundle during install.
}
