% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsdrtools-classes.R
\name{tsdrtools_install}
\alias{tsdrtools_install}
\title{Constructor for tsdrtools_install}
\usage{
tsdrtools_install(x)
}
\arguments{
\item{x}{list to with status, package, version and install log}
}
\description{
Constructor for tsdrtools_install
}
