% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsdrtools-classes.R
\name{tsdrtools_status}
\alias{tsdrtools_status}
\title{Constructor for tsdrtools_status}
\usage{
tsdrtools_status(success, pkg, version)
}
\arguments{
\item{success}{character vector of "success" or "failed"}

\item{pkg}{character vector of packages}

\item{version}{character vector of versions}
}
\description{
Constructor for tsdrtools_status
}
