% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsd_package.R
\name{tsd_package_install}
\alias{tsd_package_install}
\title{Install package on TSD from archive}
\usage{
tsd_package_install(zip_file, verbose = TRUE, ...)
}
\arguments{
\item{zip_file}{path to zipped file}

\item{verbose}{logical if status messages should be printed}

\item{...}{additional arguments to functions \code{\link[utils]{install.packages}}
and \code{\link[utils]{available.packages}}}
}
\value{
vector of successful or failed package installs
}
\description{
After downloading and preparing the
packages for install outside of TSD
using the \code{\link{tsd_package_prepare}}
function, install said packages using
this function.
}
\examples{
\dontrun{
# prepare a package first
tsd_package_prepare("devtools")

tsd_package_install("devtools.zip")
}
}
