% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsdrtools-classes.R
\name{tsdrtools_install}
\alias{tsdrtools_install}
\title{Constructor for tsdrtools_install}
\usage{
tsdrtools_install(success, error_logs)
}
\arguments{
\item{success}{data.frame of class \code{tsdrtools_status}}

\item{error_logs}{list of class \code{install_error}}
}
\description{
Constructor for tsdrtools_install
}
